backend/backend_requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlalchemy==2.1.0
pydantic==1.10.12
python-dotenv==1.0.1
